+++ src/Exceptions/InvalidOrderDataException.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace GingerPluginSdk\Exceptions;
-
-use Exception;
-
-class InvalidOrderDataException extends Exception
-{
-}